Register required as a Draft validator factory on type object
Previously "required" had no entry in the object Type's keyword registry;
its only effect (attaching RequiredPropertyValidator) was hardcoded inside
PropertyFactory::buildProperty(), bypassing the modifier/sourceKey system
every other keyword goes through, with a fragile pointer reconstructed by
string-splitting the property's own JSON pointer.

Add RequiredValidatorFactory, registered under 'required' on the object
Type right after 'properties'. It attaches RequiredPropertyValidator to
each named required property (now using the parent schema's own pointer
directly, no reconstruction needed) and fabricates a stub property for a
required name that isn't declared in 'properties'. Extract the shared
dependency-attachment logic PropertiesValidatorFactory and
RequiredValidatorFactory both need into PropertyDependencyTrait, since a
dependencies entry can target a required-but-undeclared property name.

The per-property required boolean used for nullability/attribute decisions
at construction time stays computed in PropertiesValidatorFactory, since
several modifiers (TypeCheckModifier, ConstModifier,
DefaultArrayToEmptyArrayModifier, EnumValidatorFactory) read isRequired()
synchronously while that same property is being built.
