+++ src/Model/Validator/Factory/Object/PropertyDependencyTrait.php
@@ -1,0 +1,85 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPModelGenerator\Model\Validator\Factory\Object;
+
+use PHPModelGenerator\Exception\SchemaException;
+use PHPModelGenerator\Model\Property\PropertyInterface;
+use PHPModelGenerator\Model\Schema;
+use PHPModelGenerator\Model\SchemaDefinition\JsonSchema;
+use PHPModelGenerator\Model\Validator\PropertyDependencyValidator;
+use PHPModelGenerator\Model\Validator\SchemaDependencyValidator;
+use PHPModelGenerator\PropertyProcessor\Decorator\SchemaNamespaceTransferDecorator;
+use PHPModelGenerator\SchemaProcessor\SchemaProcessor;
+
+/**
+ * Attaches a property or schema dependency (the 'dependencies' keyword) to a property.
+ * Shared by PropertiesValidatorFactory and RequiredValidatorFactory, as a property's
+ * 'dependencies' entry applies regardless of whether the property was declared in 'properties'
+ * or only exists because it is listed in 'required'.
+ */
+trait PropertyDependencyTrait
+{
+    /**
+     * @throws SchemaException
+     */
+    private function addDependencyValidator(
+        PropertyInterface $property,
+        JsonSchema $dependencyJsonSchema,
+        SchemaProcessor $schemaProcessor,
+        Schema $schema,
+    ): void {
+        $propertyDependency = true;
+
+        foreach ($dependencyJsonSchema->getJson() as $index => $dependency) {
+            if (!is_int($index) || !is_string($dependency)) {
+                $propertyDependency = false;
+                break;
+            }
+        }
+
+        $dependencyPointer = $dependencyJsonSchema->getPointer();
+
+        if ($propertyDependency) {
+            $property->addValidator(
+                (new PropertyDependencyValidator($property, $dependencyJsonSchema->getJson()))
+                    ->withJsonPointer($dependencyPointer),
+            );
+
+            return;
+        }
+
+        $json = $dependencyJsonSchema->getJson();
+        if (!isset($json['type'])) {
+            $dependencyJsonSchema = $dependencyJsonSchema->withJson($json + ['type' => 'object']);
+        }
+
+        $dependencySchema = $schemaProcessor->processSchema(
+            $dependencyJsonSchema,
+            $schema->getClassPath(),
+            "{$schema->getClassName()}_{$property->getName()}_Dependency",
+            $schema->getSchemaDictionary(),
+        );
+
+        $property->addValidator(
+            (new SchemaDependencyValidator($schemaProcessor, $property, $dependencySchema))
+                ->withJsonPointer($dependencyPointer),
+        );
+        $schema->addNamespaceTransferDecorator(new SchemaNamespaceTransferDecorator($dependencySchema));
+
+        $this->transferDependentPropertiesToBaseSchema($dependencySchema, $schema);
+    }
+
+    private function transferDependentPropertiesToBaseSchema(Schema $dependencySchema, Schema $schema): void
+    {
+        foreach ($dependencySchema->getProperties() as $dependencyProperty) {
+            $schema->addProperty(
+                (clone $dependencyProperty)
+                    ->setRequired(false)
+                    ->setType(null)
+                    ->filterValidators(static fn(): bool => false),
+            );
+        }
+    }
+}
